+++ Runtime/Core/ServiceLocator.cs
@@ -3,22 +3,22 @@ using System.Collections.Generic;
 
 namespace GameWarriors.DependencyInjection
 {
-    public static class ServiceLocator
+    public class ServiceLocator
     {
         private static Dictionary<Type, object> _serviceTable;
 
-        static ServiceLocator()
+        public ServiceLocator(int size)
         {
-            _serviceTable = new Dictionary<Type, object>(25);
+            _serviceTable = new Dictionary<Type, object>(size);
         }
 
-        public static bool Register<T>(T input)
+        public bool Register<T>(T input)
         {
             _serviceTable.Add(typeof(T), input);
             return true;
         }
 
-        public static void Register(Type injectType, object serviceObject)
+        public void Register(Type injectType, object serviceObject)
         {
             if (_serviceTable.ContainsKey(injectType))
             {
@@ -28,7 +28,7 @@ namespace GameWarriors.DependencyInjection
             _serviceTable.Add(injectType, serviceObject);
         }
 
-        public static object Resolve(Type serviceType)
+        public object Resolve(Type serviceType)
         {
             if (_serviceTable.TryGetValue(serviceType, out var service))
             {
@@ -38,7 +38,7 @@ namespace GameWarriors.DependencyInjection
         }
 
 
-        public static T Resolve<T>()
+        public T Resolve<T>()
         {
             if (_serviceTable.TryGetValue(typeof(T), out var service))
             {
@@ -47,7 +47,5 @@ namespace GameWarriors.DependencyInjection
 
             return default;
         }
-
-
     }
 }