+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-  version : 0.2.0
-  Change and improvement on service collection loading pipeline, Inject properties now has value inside WaitForLoading() method
+  version : 0.4.0
+  Change the service locator inside srevice provider class from static to instance object class
 
   version : 0.3.0
   Add method call after wait for loading step in service collection. The method name can be customize and can has optional parameter for injection
+
+  version : 0.2.0
+  Change and improvement on service collection loading pipeline, Inject properties now has value inside WaitForLoading() method